+++ src/google/protobuf/compiler/objectivec/message.h
@@ -8,7 +8,6 @@
 #ifndef GOOGLE_PROTOBUF_COMPILER_OBJECTIVEC_MESSAGE_H__
 #define GOOGLE_PROTOBUF_COMPILER_OBJECTIVEC_MESSAGE_H__
 
-#include <cstddef>
 #include <memory>
 #include <string>
 #include <vector>
@@ -62,7 +61,7 @@ class MessageGenerator {
   const std::string deprecated_attribute_;
   std::vector<const ExtensionGenerator*> extension_generators_;
   std::vector<std::unique_ptr<OneofGenerator>> oneof_generators_;
-  size_t sizeof_has_storage_;
+  int sizeof_has_storage_;
 };
 
 }  // namespace objectivec
